exposed async api as well
